requirements_dev.txt
requests==2.26.0
typing-extensions==3.10.0.2
pytest-dotenv==0.5.2
pytest==6.2.5
black==21.11b1
isort==5.10.1
flake8==4.0.1
pre-commit==2.15.0
